LayoutTests/fast/dom/shadow/shadow-root-new.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p>This test checks the object 'new WebKitShadowRoot(host)' and internals.shadowRoot(div) are the same.</p>
<pre id="console"></pre>

<script>
function doTest() {
    if (!window.internals) {
        debug('This test runs on DRT only');
        return;
    }

    div = document.createElement('div');
    // we should call new WebKitShadowRoot first.
    root1 = new WebKitShadowRoot(div);
    root2 = internals.shadowRoot(div);

    shouldBe('root1', 'root2');
}

doTest();
var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
